app/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-oauth-toolkit
django-cors-headers
celery==5.2.7
git+https://github.com/thepylot/djongo.git
mongoengine==0.20.0
pylint-mongoengine==0.4.0
pymongo==3.11.0
psycopg2-binary==2.8.5
redis==3.5.3
Faker
requests-oauthlib==1.3.1
django_otp